DO NOT DOWNLOAD. DOWNLOAD LATEST STABLE FROM "DOWNLOADS" TAB Further performance optimization
--HG--
extra : convert_revision : svn%3A712752d5-f020-ad4b-8268-b7cbb332fca5/trunk%401035

+++ NopCommerceStore/Administration/Modules/PaymentMethodsFilterControl.ascx.cs
@@ -248,7 +248,10 @@ namespace NopSolutions.NopCommerce.Web.Administration.Modules
                     }
                     else
                     {
-                        PaymentMethodManager.DeletePaymentMethodCountryMapping(paymentMethod.PaymentMethodId, countryId);
+                        if (PaymentMethodManager.DoesPaymentMethodCountryMappingExist(paymentMethod.PaymentMethodId, countryId))
+                        {
+                            PaymentMethodManager.DeletePaymentMethodCountryMapping(paymentMethod.PaymentMethodId, countryId);
+                        }
                     }
                 }
             }
